requirements.txt
Flask == 3.0.1
requests == 2.31.0
python-dotenv == 1.0.1
openai == 1.11.1
sqlalchemy == 2.0.27
pyodbc == 5.1.0